[refactor] moved sitewise api func to separate module
