+++ pkg/sitewise/api/describe_asset_property.go
@@ -1,4 +1,4 @@
-package sitewise
+package api
 
 import (
 	"context"
