+++ pkg/sitewise/api/list_asset_models.go
@@ -1,4 +1,4 @@
-package sitewise
+package api
 
 import (
 	"context"
