+++ pkg/sitewise/api/list_associated_assets.go
@@ -1,4 +1,4 @@
-package sitewise
+package api
 
 import (
 	"context"
