+++ pkg/sitewise/api/property_aggregate.go
@@ -1,4 +1,4 @@
-package sitewise
+package api
 
 import (
 	"context"
