+++ pkg/sitewise/api/property_history.go
@@ -1,4 +1,4 @@
-package sitewise
+package api
 
 import (
 	"context"
